Rewrite main window handling (temp)
