--- src/main/mainWindow/index.js
@@ -1,93 +1,0 @@
-import { app, BrowserWindow } from 'electron';
-
-import { WindowStateHandler } from './state';
-
-
-let mainWindow = null;
-
-let state = {
-	hideOnClose: false,
-};
-
-const setState = (partialState) => {
-	state = {
-		...state,
-		...partialState,
-	};
-};
-
-async function attachWindowStateHandling(mainWindow) {
-	const windowStateHandler = new WindowStateHandler(mainWindow, 'main');
-	await windowStateHandler.load();
-	await new Promise((resolve) => mainWindow.once('ready-to-show', resolve));
-	windowStateHandler.apply();
-
-	const exitFullscreen = () => new Promise((resolve) => {
-		if (mainWindow.isFullScreen()) {
-			mainWindow.once('leave-full-screen', resolve);
-			mainWindow.setFullScreen(false);
-			return;
-		}
-		resolve();
-	});
-
-	const close = () => {
-		mainWindow.blur();
-
-		if (process.platform === 'darwin' || state.hideOnClose) {
-			mainWindow.hide();
-		} else if (process.platform === 'win32') {
-			mainWindow.minimize();
-		} else {
-			app.quit();
-		}
-	};
-
-	app.on('activate', () => mainWindow && mainWindow.show());
-	app.on('before-quit', () => {
-		mainWindow = null;
-		windowStateHandler.save();
-	});
-
-	mainWindow.on('resize', () => windowStateHandler.fetchAndSave());
-	mainWindow.on('move', () => windowStateHandler.fetchAndSave());
-	mainWindow.on('show', () => windowStateHandler.fetchAndSave());
-	mainWindow.on('close', async (event) => {
-		if (!mainWindow) {
-			return;
-		}
-
-		event.preventDefault();
-		await exitFullscreen();
-		close();
-		windowStateHandler.fetchAndSave();
-	});
-
-	mainWindow.on('set-state', setState);
-}
-
-export async function createMainWindow() {
-	mainWindow = new BrowserWindow({
-		width: 1000,
-		height: 600,
-		minWidth: 600,
-		minHeight: 400,
-		titleBarStyle: 'hidden',
-		show: false,
-		webPreferences: {
-			webviewTag: true,
-			nodeIntegration: true,
-		},
-	});
-	attachWindowStateHandling(mainWindow);
-
-	mainWindow.webContents.on('will-attach-webview', (event, webPreferences) => {
-		delete webPreferences.enableBlinkFeatures;
-	});
-
-	mainWindow.loadFile(`${ app.getAppPath() }/app/public/app.html`);
-
-	if (process.env.NODE_ENV === 'development') {
-		mainWindow.webContents.openDevTools();
-	}
-}
